Making TypeSystem a class-type, instead of a base protocol
