+++ Sources/SwiftRewriterLib/TypeSystem/TypeCategory.swift
@@ -1,0 +1,11 @@
+public enum TypeCategory {
+    case integer
+    case float
+    case boolean
+    case void
+    case `struct`
+    case `protocol`
+    case `enum`
+    case `class`
+    case unknown
+}
